requirements.txt
streamlit
python-dotenv
requests
google-cloud-bigquery
google-api-core
pandas
plotly
